Add files via upload
The physiological signals were acquired by the ESI NeuroScan system with a 1000 Hz sampling rate and a preset band-pass filter between 0.1 Hz and 40 Hz.  Channel 1 and 2 signals are from two horizontal EOG (HEOG) electrodes, which locate at the outer canthus of the left eye and the right eye. Besides, Channel 3 and 4 signals are from two vertical EOG (VEOG) electrodes, which are placed at the upper and lower sides of the left eye.